pyproject.toml
[tool.poetry]
name = "multi-tool"
version = "0.1.0"
description = "The Multi-Purpose API Toolkit provides a robust and cohesive collection of APIs designed to facilitate a wide array of common yet pivotal tasks for developers. This toolkit consolidates diverse functionalities into a singular endpoint, simplifying the process of integrating multiple third-party services. Key offerings include:"
authors = ["AutoGPT <info@agpt.co>"]
readme = "README.md"

[tool.poetry.dependencies]
python = ">=3.11"
pillow = "^9.2.0"
beautifulsoup4 = "^4.9.3"
email-validator = "^1.1.3"
fastapi = "*"
google-cloud-texttospeech = "^2.10.0"
httpx = "^0.23.0"
prisma = "*"
pydantic = "*"
python-barcode = "*"
python-dateutil = "^2.8.2"
pytz = "^2022.1"
qrcode = "*"
uvicorn = "*"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
